+++ CHANGELOG.rst
@@ -6,6 +6,9 @@ Changelog
 v0.4 (work in progress)
 -----------------------
 
+- Disabling ``mv_store`` feature of H2.
+- Fixing various bugs in ``H2ScoreDistributionReader`` and ``H2ScoreDistributionWriter``.
+
 ----
 v0.3
 ----